refatora Jenkinsfile: melhora comandos para parar e remover containers, adiciona verificação do container Flask antes de executar testes

+++ Jenkinsfile
@@ -7,7 +7,8 @@ pipeline {
                 script {
                     // Parar e remover containers existentes, se houver
                     sh '''
-                    docker stop $(docker ps -aq) && docker rm $(docker ps -aq)
+                    docker ps -aq | xargs -r docker stop
+                    docker ps -aq | xargs -r docker rm
                     docker-compose down || true
                     docker system prune -f || true
                     '''
@@ -21,18 +22,30 @@ pipeline {
                 sh 'docker-compose build --no-cache'
             }
         }
-
+        
         stage('Testar Aplicação') {
             steps {
                 echo "=== Executando testes ==="
                 script {
                     sh 'docker-compose up -d mariadb flask'
                     sh 'sleep 10' // Tempo para inicializar os serviços
+                    
+                    // Verifica se o container Flask está ativo
+                    sh '''
+                    CONTAINER_ID=$(docker ps -qf "name=flask")
+                    if [ -z "$CONTAINER_ID" ]; then
+                        echo "Erro: o container Flask não está em execução!"
+                        exit 1
+                    fi
+                    '''
+                    
+                    // Executa os testes
                     sh 'docker exec $(docker ps -qf "name=flask") pytest /app/tests'
                     sh 'docker-compose down'
                 }
             }
         }
+
 
         
         stage('Run Containers') {
